Refactoring: separated project for tests.
