+++ cs/TagsCloudVisualizationTest/CircularCloudLayouter_Should.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using NUnit.Framework;
+using TagsCloudVisualization;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualizationTest
 {
     class CircularCloudLayouter_Should
     {
