+++ cs/TagsCloudVisualizationTest/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MSTest.TestAdapter" version="1.2.1" targetFramework="net461" />
+  <package id="MSTest.TestFramework" version="1.2.1" targetFramework="net461" />
+  <package id="NUnit" version="3.11.0" targetFramework="net461" />
+</packages>
